agent/requirements.txt
# ai-interviewer-meet/agent/requirements.
livekit-api
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
python-dotenv
livekit-rtc